--- src/app/my-blending/MyBlendingPage.module.scss
@@ -1,11 +1,0 @@
-.main {
-  margin-bottom: 24.4rem;
-
-  @include tablet {
-    margin-bottom: 45.2rem;
-  }
-
-  @include mobile {
-    margin-bottom: 22.367rem;
-  }
-}
