--- src/components/myBlending/TeaCardList/TeaCardList.module.scss
@@ -1,4 +1,0 @@
-.cardList {
-  @include flexSort(row, null, null, 1rem 2rem);
-  flex-wrap: wrap;
-}
