+++ src/components/page-layout/myBlendingLayout/MyBlendingLayout.module.scss
@@ -1,0 +1,25 @@
+.layout {
+  width: 100%;
+  padding-right: 4rem;
+  padding-left: 4rem;
+  margin-bottom: 24.4rem;
+
+  .container {
+    margin: 0 auto;
+    width: 100%;
+    min-width: 33.5rem;
+    max-width: 120rem;
+  }
+
+  @include tablet {
+    margin-bottom: 45.2rem;
+    padding-right: 2rem;
+    padding-left: 2rem;
+  }
+
+  @include mobile {
+    margin-bottom: 22.367rem;
+    padding-right: 1.7rem;
+    padding-left: 1.7rem;
+  }
+}
