+++ src/components/page-layout/myBlendingLayout/components/FixedBottomBar/FixedBottomBar.module.scss
