+++ src/components/page-layout/myBlendingLayout/components/SortingMenu/SortingList/index.tsx
@@ -1,0 +1,8 @@
+import styles from '@/components/page-layout/myBlendingLayout/components/SortingMenu/SortingList/SortingList.module.scss';
+import classNames from 'classnames/bind';
+
+const cn = classNames.bind(styles);
+
+export default function SortingList() {
+  return <ul>{}</ul>;
+}
