+++ src/components/page-layout/myBlendingLayout/components/SortingMenu/SortingMenu.module.scss
@@ -1,0 +1,20 @@
+.container {
+  @include flexSort(column, null, null, 2.8rem);
+  padding: 4rem 0;
+
+  .title {
+    @include font(2.8rem, 700, 3.8rem);
+  }
+
+  .menu {
+    @include flexSort(column, null, null, 1.2rem);
+
+    .buttons {
+      @include flexSort(row, null, center, 1.6rem);
+
+      .criterion {
+        @include font(1.7rem, 600, 2.4rem);
+      }
+    }
+  }
+}
