+++ src/components/page-layout/myBlendingLayout/components/TeaCardList/TeaCard/CheckIcon/CheckIcon.module.scss
