+++ src/components/page-layout/myBlendingLayout/components/TeaCardList/TeaCardList.module.scss
@@ -1,0 +1,4 @@
+.cardList {
+  @include flexSort(row, center, null, 2rem);
+  flex-wrap: wrap;
+}
